Some Bug Fixes

+++ seamos-amigos/src/main/java/com/himansh/seamosamigos/service/CommentAndReplyService.java
@@ -7,9 +7,11 @@ import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import com.himansh.seamosamigos.dto.CommentWebModel;
 import com.himansh.seamosamigos.entity.Comments;
+import com.himansh.seamosamigos.exception.InAppException;
 import com.himansh.seamosamigos.repository.CommentRepository;
 
 @Service
@@ -19,14 +21,14 @@ public class CommentAndReplyService {
 	
 	public List<CommentWebModel> getAllComments(Integer photoId){
 		return commentRepository.getAllCommentsByPicId(photoId).stream()
-				.filter(c->c.getPhotoId()!=null && c.getPhotoId()!=0)
+//				.filter(c->c.getPhotoId()!=null && c.getPhotoId()!=0)
 				.map(CommentWebModel::toWebModel).collect(Collectors.toList());
 	}
 	
 	public CommentWebModel saveComment(CommentWebModel commentWebModel, int userId) {
 		commentWebModel.setUserId(userId);
 		if (commentWebModel.getCommentId()!=null && commentWebModel.getCommentId()!=0) {
-			commentWebModel.setPhotoId(null);
+			//commentWebModel.setPhotoId(null);
 			Optional<Comments> comment= commentRepository.findById(commentWebModel.getCommentId());
 			if (comment.isPresent()) {
 				Set<Comments> replies=comment.get().getReplies();
@@ -41,12 +43,47 @@ public class CommentAndReplyService {
 		return CommentWebModel.toWebModel(comment);
 	}
 	
-	public Integer deleteComment(Integer commentId) {
+	public Integer deleteComment(Integer commentId, Integer userId) {
 		Optional<Comments> comment= commentRepository.findById(commentId);
 		if(comment.isPresent()) {
-			commentRepository.deleteById(commentId);
-			return commentId;
+			Comments delComment= comment.get();
+			boolean flag= false;
+			if (userId == delComment.getUserId()) {
+				flag= true;
+			}
+			else {
+				Integer picId=delComment.getPhotoId();
+				System.out.println(commentRepository.findUserIdByPicId(picId).size());
+				if(commentRepository.findUserIdByPicId(picId).contains(userId))
+					flag= true;
+			}
+			if(flag) {
+				commentRepository.deleteById(commentId);
+				return commentId;
+			}
 		}
 		return 0;
+	}
+
+	public Object updateComment(CommentWebModel commentWebModel, Integer userId) throws InAppException {
+		Integer commentId= commentWebModel.getCommentId();
+		Optional<Comments> comment= commentRepository.findById(commentId);
+		if (comment.isPresent()) {
+			Comments updatesOnComent = comment.get();
+			if (userId != updatesOnComent.getUserId()) {
+				throw new InAppException("User not allowed to update");
+			}
+			//Integer likes= commentWebModel.getLikes();
+			String message= commentWebModel.getBody();
+			if(!StringUtils.isEmpty(message))
+				updatesOnComent.setBody(commentWebModel.getBody());
+			//else if(!StringUtils.isEmpty(likes+""))
+			//	updatesOnComent.setLikes(commentWebModel.getLikes());
+			else
+				throw new InAppException("comment is empty");
+			updatesOnComent= commentRepository.saveAndFlush(updatesOnComent);
+			return updatesOnComent;
+		}
+		return null;
 	}
 }
